// src/app/enum/status-messages.enum.ts
export enum StatusMessages {
    MarkTaskAsCompleted = 'Task marked as completed',
    TasksFetchedSuccessfully = 'Tasks fetched successfully',
    TaskCreatedSuccessfully = 'Task created successfully',
    TaskUpdatedSuccessfully = 'Task updated successfully',
    NewTaskAddedSuccessfully =  'New task added successfully',
    TaskDeletedSuccessfully = 'Task deleted successfully',

    TaskCompleted = 'Completed tasks cannot be moved!',

    RequireFormFields = 'Require fields are missing',

    InvalidCredential = 'Invalid credentials please try again.',

    Error = 'An error occurred please try again later.',

}